backend/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
psycopg2-binary==2.9.10
pydicom==3.0.1
python-dotenv==1.1.1
django-cors-headers==4.7.0
dj-database-url==2.2.0
gunicorn==22.0.0

# Development dependencies (optional)
django-debug-toolbar==4.2.0
pytest-django==4.8.0
black==24.4.2
flake8==7.0.0 